Fix printf warning when long long is 128-bit.

+++ rflags.c
@@ -5,7 +5,7 @@ int main()
 {
 	uint64_t rflags;
 	__asm__ __volatile__ ("pushf; popq %0" : "=g" (rflags) ::"memory");
-	printf("rFLAGS: %016llx\n", rflags);
+	printf("rFLAGS: %016llx\n", (unsigned long long)rflags);
 #define BIT(name, number) \
 	printf("\t%s: %d\n", name, rflags & (1 << number) ? 1 : 0)
